+++ packages/cortex-react/src/provider/index.ts
@@ -3,3 +3,4 @@ export * from './SnackbarProvider';
 export * from './MenubarContext';
 export * from './useMenubar';
 export * from './CalendarProvider';
+export * from './ManagerContext';